==> Codes/14_Comparisons.ipynb <==
"""
Find the files in a particular AFE framework folder
not in Base. Used to identify which dataframes were not processed by a given framework

"""

import os

dir_basic = r'C:\Users\25644574\OneDrive - UTS\Desktop\Updated Codes New\Data\BASIC'
dir_AFE = r'C:\Users\25644574\OneDrive - UTS\Desktop\Updated Codes New\Data\AutoFeat'

# Get a list of file names in both directories
files_basic = set(os.listdir(dir_basic))
files_AFE = set(os.listdir(dir_AFE))

# Find files that are in BASIC but not in AFE Folder
unique_files = files_basic - files_AFE

# Print the unique files
print("Files in BASIC not in AFE Folder:")
for file in unique_files:
    print(file)
