==> Codes/15_Renaming_Models.ipynb <==
Import the libraries

import pickle
import os

Getting the results

# Paths to the pickle files
Base_path = '../Results/Base_Results.pkl'
BigFeat_path = '../Results/BigFeat_results.pkl'
AutoFeat_path = '../Results/AutoFeat_results.pkl'
FT_path = '../Results/FT_Results.pkl'
Open_path = '../Results/OpenFE_Results.pkl'
SAFE_path = '../Results/SAFE_Results.pkl'

# Load the data
with open(Base_path, 'rb') as f:
    Base = pickle.load(f)

with open(BigFeat_path, 'rb') as f:
    BigFeat = pickle.load(f)

with open(AutoFeat_path, 'rb') as f:
    AutoFeat = pickle.load(f)

with open(FT_path, 'rb') as f:
    FeatureTools = pickle.load(f)

with open(Open_path, 'rb') as f:
    OpenFE = pickle.load(f)

with open(SAFE_path, 'rb') as f:
    SAFE = pickle.load(f)

Renaming Models

# Dictionary to map old names to new abbreviations
rename_dict = {
    'AdaBoost Classification': 'AB',
    'Decision Tree Classification': 'DT',
    'Extremely Randomized Trees Classification': 'ET',
    'Gradient Boosting Classification': 'GB',
    'KNN Classification': 'KNN',
    'LightGBM Classification': 'LGBM',
    'Logistic Regression': 'LR',
    'MLP Classification': 'MLP',
    'Random Forest Classification': 'RF',
    'SVM Classification': 'SVM',
    'XGBoost Classification': 'XGB',
    'AdaBoost Classification Multi-Class': 'AB',
    'Decision Tree Classification Multi-Class': 'DT',
    'Extremely Randomized Trees Classification Multi-Class': 'ET',
    'Gradient Boosting Classification Multi-Class': 'GB',
    'KNN Classification Multi-Class': 'KNN',
    'LightGBM Classification Multi-Class': 'LGBM',
    'Logistic Regression Multi-Class': 'LR',
    'MLP Classification Multi-Class': 'MLP',
    'Random Forest Classification Multi-Class': 'RF',
    'SVM Classification Multi-Class': 'SVM',
    'XGBoost Classification Multi-Class': 'XGB'
}

# Replace the values based on the dictionary
Base['classification_df']['Model'] = Base['classification_df']['Model'].replace(rename_dict)
Base['classification_all']['Model'] = Base['classification_all']['Model'].replace(rename_dict)
Base['classification_df_std']['Model'] = Base['classification_df_std']['Model'].replace(rename_dict)

BigFeat['classification_df']['Model'] = BigFeat['classification_df']['Model'].replace(rename_dict)
BigFeat['classification_all']['Model'] = BigFeat['classification_all']['Model'].replace(rename_dict)
BigFeat['classification_df_std']['Model'] = BigFeat['classification_df_std']['Model'].replace(rename_dict)

AutoFeat['classification_df']['Model'] = AutoFeat['classification_df']['Model'].replace(rename_dict)
AutoFeat['classification_all']['Model'] = AutoFeat['classification_all']['Model'].replace(rename_dict)
AutoFeat['classification_df_std']['Model'] = AutoFeat['classification_df_std']['Model'].replace(rename_dict)

FeatureTools['classification_df']['Model'] = FeatureTools['classification_df']['Model'].replace(rename_dict)
FeatureTools['classification_all']['Model'] = FeatureTools['classification_all']['Model'].replace(rename_dict)
FeatureTools['classification_df_std']['Model'] = FeatureTools['classification_df_std']['Model'].replace(rename_dict)

OpenFE['classification_df']['Model'] = OpenFE['classification_df']['Model'].replace(rename_dict)
OpenFE['classification_all']['Model'] = OpenFE['classification_all']['Model'].replace(rename_dict)
OpenFE['classification_df_std']['Model'] = OpenFE['classification_df_std']['Model'].replace(rename_dict)

SAFE['classification_df']['Model'] = SAFE['classification_df']['Model'].replace(rename_dict)
SAFE['classification_all']['Model'] = SAFE['classification_all']['Model'].replace(rename_dict)
SAFE['classification_df_std']['Model'] = SAFE['classification_df_std']['Model'].replace(rename_dict)

# Dictionary to map old names to new abbreviations
rename_dict = {
    'Linear Regression': 'LR',
    'KNN Regression': 'KNN',
    'SVM Regression': 'SVM',
    'Random Forest Regression': 'RF',
    'AdaBoost Regression': 'AB',
    'MLP Regression': 'MLP',
    'Decision Tree Regression': 'DT',
    'Extremely Randomized Trees Regression': 'ET',
    'Gradient Boosting Regression': 'GB',
    'LightGBM Regression': 'LGBM',
    'XGBoost Regression': 'XGB'
}

# Replace the values based on the dictionary
Base['regression_df']['Model'] = Base['regression_df']['Model'].replace(rename_dict)
Base['regression_all']['Model'] = Base['regression_all']['Model'].replace(rename_dict)
Base['regression_df_std']['Model'] = Base['regression_df_std']['Model'].replace(rename_dict)

AutoFeat['regression_df']['Model'] = AutoFeat['regression_df']['Model'].replace(rename_dict)
AutoFeat['regression_all']['Model'] = AutoFeat['regression_all']['Model'].replace(rename_dict)
AutoFeat['regression_df_std']['Model'] = AutoFeat['regression_df_std']['Model'].replace(rename_dict)

FeatureTools['regression_df']['Model'] = FeatureTools['regression_df']['Model'].replace(rename_dict)
FeatureTools['regression_all']['Model'] = FeatureTools['regression_all']['Model'].replace(rename_dict)
FeatureTools['regression_df_std']['Model'] = FeatureTools['regression_df_std']['Model'].replace(rename_dict)

OpenFE['regression_df']['Model'] = OpenFE['regression_df']['Model'].replace(rename_dict)
OpenFE['regression_all']['Model'] = OpenFE['regression_all']['Model'].replace(rename_dict)
OpenFE['regression_df_std']['Model'] = OpenFE['regression_df_std']['Model'].replace(rename_dict)

SAFE['regression_df']['Model'] = SAFE['regression_df']['Model'].replace(rename_dict)
SAFE['regression_all']['Model'] = SAFE['regression_all']['Model'].replace(rename_dict)
SAFE['regression_df_std']['Model'] = SAFE['regression_df_std']['Model'].replace(rename_dict)

Saving the results

#Save the result
pickle.dump(Base, open("../Results/Base_Results.pkl", "wb"))
pickle.dump(BigFeat, open("../Results/BigFeat_Results.pkl", "wb"))
pickle.dump(OpenFE, open("../Results/OpenFE_Results.pkl", "wb"))
pickle.dump(SAFE, open("../Results/SAFE_Results.pkl", "wb"))
pickle.dump(AutoFeat, open("../Results/AutoFeat_Results.pkl", "wb"))
pickle.dump(FeatureTools, open("../Results/FT_Results.pkl", "wb"))
